# performance improvements for stickers