Adding support for stickers to the Ubuntu Telegram app, I noticed that views on a sticker model are very slow. This is also noticeable with Cutegram on my laptop. When scrolling something, it stutters whenever a new sticker is created. I tracked this down to the file cache, which iterates over all files in the directory. On my bq phone this takes about 15 ms per cache lookup. In combination with the fact that the stickerfilemanager checks the cache twice, once for the file and once for the thumbnail, this takes a bit more than 30ms to look up one sticker. 

This diff doesn't loop over all the files in the cache when we precisely know the file name already, as it is the case for stickers, instead it just returns the filename immediately, saving the whole 30ms per sticker load operation.

For all other types (e.g. regular images) the cache is still iterated as before, however, instead of looping the whole directory in the manual Q_FOREACH, this sets a filter on QDir::entryList. This gets the lookup from 15ms down to 11ms, so this patch also slightly improves other files. I would still suggest to remove the looping over cache files completely and rely on more predictable file names. This would improve the overall scrolling and loading experience in the whole Cutegram application and its derivates quite a bit (seeing that such cache lookups are done *very* often throughout the app usage).


Doing this, I also found an issue that sticker files were cached as "basename.webp.webp". In order to make the above fix work, I added a check in uploadGetFile_slt to prevent adding the suffix a second time if it is already there. I'm not sure if that's really the best way, or if the whole logic should be reworked a bit, but I didn't want to start refactoring too much in this current task (adding sticker support).

Hope this helps.